+++ app/src/main/AndroidManifest.xml
@@ -35,6 +35,7 @@
         </activity>
         <activity android:name=".StartActivity" />
         <activity android:name=".CameraActivity" />
+        <activity android:name=".AlbumActivity" />
     </application>
 
 </manifest>
